# static/openapi.yaml
openapi: 3.0.3
info:
  title: SodhiCable MES v4.0 API
  description: |
    REST API for the SodhiCable Manufacturing Execution System.

    Aligned with **MESA-11** (Manufacturing Enterprise Solutions Association) functional
    model and the **ISA-95** (IEC 62264) enterprise-control integration standard.

    The API spans ISA-95 Levels 1-4:
    - **Level 1-2** - Process control, sensors, equipment monitoring
    - **Level 3** - MES core: work orders, scheduling, quality, traceability, performance, labor
    - **Level 4** - ERP integration: S&OP, purchase orders, order-to-cash, ISA-95 messaging
    - **Cross-Level** - AI/ML intelligence layer (proposed MESA F12)

    All responses use JSON. Paginated endpoints return an envelope with
    `{data, page, per_page, total}`.
  version: 4.0.0
  contact:
    name: Stephen Eacuello
    email: seacuello@uri.edu
  license:
    name: Academic Use
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/process/deviations:
    get:
      tags: ["Level 1-2 - Process Control"]
      summary: Process deviations (alarms)
      description: |
        Returns process deviation records. Filterable by work center, severity,
        and resolution status. Limited to 200 records.
      operationId: getProcessDeviations
      parameters:
        - name: wc_id
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
            enum: [Minor, Warning, Major, Critical]
        - name: resolved
          in: query
          schema:
            type: string
            enum: ["0", "1"]
          description: "0 = unresolved, 1 = resolved"
      responses:
        "200":
          description: Deviation records
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    deviation_id:
                      type: integer
                    wc_id:
                      type: string
                    parameter_name:
                      type: string
                    severity:
                      type: string
                    deviation_value:
                      type: number
                    setpoint_value:
                      type: number
                    resolved:
                      type: integer
                    timestamp:
                      type: string
...